Make changes to database schema
Updated the schema to reflect the changes in how we are going to implement the message data relative to a user's location. JSON data types were added to Messages, Events, and User. Some other minor additions and removals of columns to increase the capability of our User Experience are also implemented.